new /queue command

+++ utils/musicPlayer.js
@@ -2,11 +2,11 @@ const { EmbedBuilder, ButtonBuilder, ActionRowBuilder, ButtonStyle } = require("
 const wait = require("util").promisify(setTimeout);
 const logger = require("../utils/logger");
 
-let msg = null;
+let msg = null; 
 
 module.exports = {
     trackStart: async (client, queue, track) => {
-        if (msg) msg.delete();
+        if (msg != null) return; // Prevents multiple messages from being sent
         //console.log(queue)
         //console.log(track)
         const channel = queue.metadata.channel;
@@ -34,7 +34,7 @@ module.exports = {
         const player = new EmbedBuilder()
         .setTitle(`🎧 Now Playing ${queue.dispatcher.channel.name ? `in ${queue.dispatcher.channel.name}` : ""}`)
         .setAuthor({ name: `Requested by ${requestedBy.username}`, iconURL: requestedBy.displayAvatarURL({dynamic: true}) })
-        .setDescription(`${desc}\n\n${track.isStream ? `🔴 LIVE` : `🔘 ${queue.node.createProgressBar()} 🔘`}\n\n${queue.tracks.length < 0 ? `Up Next: [${queue.tracks[0].title}](${queue.tracks[0].url})\n By **${queue.tracks[0].author}**` : ``}`)
+        .setDescription(`${desc}\n\n${track.isStream ? `🔴 LIVE` : `🔘 ${queue.node.createProgressBar()} 🔘`}\n\n${queue.tracks.length > 0 ? `Up Next: [${queue.tracks[0].title}](${queue.tracks[0].url})\n By **${queue.tracks[0].author}**` : ``}`)
         .setThumbnail(track.thumbnail)
         .setColor(0x00AE86)
         .setFooter({ text: `Meme Cultist | Version ${require('../package.json').version}`, iconURL: client.user.displayAvatarURL({dynamic: true}) })
@@ -44,21 +44,24 @@ module.exports = {
 
         const interval = setInterval(async () => {
             if (!queue.node.isPlaying() || queue.node.isPaused() || track.isStream) return clearInterval(interval);
-            player.setDescription(`${desc}\n\n${track.isStream ? `🔴 LIVE` : `🔘 ${queue.node.createProgressBar()} 🔘`}${queue.tracks.length > 0 ? `\n\n Up Next: [${queue.tracks[0].title}](${queue.tracks[0].url})\n By **${queue.tracks[0].author}**` : ""}`)
+            player.setDescription(`${desc}\n\n${track.isStream ? `🔴 LIVE` : `🔘 ${queue.node.createProgressBar()} 🔘`}${queue.tracks.length > 0 ? `\n\n Up Next: [${queue.tracks[0].title}](${queue.tracks[0].url})\n By **${queue.tracks[0].author}**` : ``}`)
             await msg.edit({embeds: [player], components: [row]});
         }, 1000);
 
-        const filter = i => {
-            logger.log(`${i.member.voice.channelId} === ${queue.dispatcher.channel.id} = ${i.member.voice.channelId === queue.dispatcher.channel.id}`)
-            return i.member.voice.channelId === queue.dispatcher.channel.id;
-        }
+        const filter = i => i.member.voice.channelId === queue.dispatcher.channel.id;
         const collector = await channel.createMessageComponentCollector({ filter, time: (track.durationMS - queue.node.getTrackPosition()) });
     
         collector.on('collect', async i => {
-            if (!filter) return; 
+            if (!filter) return await i.reply({ content: `Join the bot's channel to use these buttons!`, ephemeral: true }); 
             logger.log(`${i.member.user.username} pressed ${i.customId}`);
             if (i.customId === "pause") {
-                logger.log(queue.node.isPlaying())
+                queue.node.isPaused() ? await queue.node.resume() : await queue.node.pause();
+                await collector.resetTimer({ time: 300000 }); // 5 minutes to respond
+                player.setTitle(queue.node.isPaused() ? `⏸️ Song Paused` : `🎧 Now Playing ${queue.dispatcher.channel.name ? `in ${queue.dispatcher.channel.name}` : ""}`);
+                row.components[0].setLabel(queue.node.isPaused() ? "Resume" : "Pause").setEmoji(queue.node.isPaused() ? "▶️" : "⏸️");
+                row.components[1].setDisabled(queue.node.isPaused());
+                await i.update({embeds: [player], components: [row]});
+                /*
                 if (!queue.node.isPaused()) {
                     logger.log("paused");
                     await queue.node.pause();
@@ -73,6 +76,7 @@ module.exports = {
                     await msg.delete();
                     return await collector.stop();
                 }
+                */
             } else if (i.customId === "skip") {
                 try {
                     if (!queue.node.isPlaying() || queue.node.isPaused()) {
@@ -80,7 +84,7 @@ module.exports = {
                         return await i.reply({ content: `Unpause before trying to skip. Too lazy to fix this bug for now.`, ephemeral: true });
                     }
                     await queue.node.skip();
-                    await msg.delete();
+                    if (msg != null) await msg.delete();
                     await collector.stop();
                 } catch (e) {
                     logger.error(e);
@@ -88,7 +92,7 @@ module.exports = {
             } else if (i.customId === "stop") {
                 try {
                     await queue.delete();
-                    await msg.delete();
+                    if (msg != null) await msg.delete();
                     return await collector.stop();
                 } catch (e) {
                     logger.error(e);
@@ -104,20 +108,27 @@ module.exports = {
                     wait(30000).then(async () => {
                         if (queue.node.isPaused()) {
                             await queue.delete();
-                            await msg.delete();
-                            await reply.delete();
+                            if (msg != null) msg.delete();
+                            if (reply) await reply.delete();
                         } else {
-                            await reply.delete();
+                            if (reply) await reply.delete();
                         }
                     });
-                } else {
-                    if (msg) msg.delete();
-                }
+                } 
             } 
         });
     },
     trackEnd: async (client, queue, track) => {
-        if (msg) msg.delete();
+        if (msg != null) msg.delete();
+        msg = null;
     },
-
+    queueString: (tracks) => {
+        let result = tracks.map((track, i) => `**${i + 1}.** [${track.title}](${track.url}) by **${track.author}** - ${track.duration}`).join("\n");
+        if (result.length > 3584) {
+            result = result.substring(0, 3584);
+            result = result.substring(0, result.lastIndexOf("\n"));
+            result += "\n...";
+        }
+        return result;
+    }
 };
